Finished seed with success

+++ EndavaEvents/routes/web.php
@@ -35,15 +35,15 @@ Route::put('/profile/{profile}', 'ProfileController@updateProfile')->middleware(
 
 //Admin
 
-Route::get('/admins/users','AdminController@index')->middleware('admin')->name('admins.users');
-Route::get('/admins/events','AdminController@viewEvents')->middleware('admin')->name('admins.events');
-Route::get('/admins/events/{event}','AdminController@showEvent')->middleware('admin')->name('admins.show');
-Route::get('/admins/report','AdminController@generateReport')->middleware('admin')->name('admins.report');
+Route::get('/admins/users','AdminController@index')->middleware('admin')->middleware('auth')->name('admins.users');
+Route::get('/admins/events','AdminController@viewEvents')->middleware('admin')->middleware('auth')->name('admins.events');
+Route::get('/admins/events/{event}','AdminController@showEvent')->middleware('admin')->middleware('auth')->name('admins.show');
+Route::get('/admins/report','AdminController@generateReport')->middleware('admin')->middleware('auth')->name('admins.report');
 
-Route::post('/admins/report/do','AdminController@doReport')->middleware('admin')->name('admins.doreport');
+Route::post('/admins/report/do','AdminController@doReport')->middleware('admin')->middleware('auth')->name('admins.doreport');
 
-Route::put('/admins/do/{user}','AdminController@doAdmin')->middleware('admin')->name('admins.doadmin');
-Route::put('/admins/remove/{user}','AdminController@removeAdmin')->middleware('admin')->name('admins.removeadmin');
+Route::put('/admins/do/{user}','AdminController@doAdmin')->middleware('admin')->middleware('auth')->name('admins.doadmin');
+Route::put('/admins/remove/{user}','AdminController@removeAdmin')->middleware('admin')->middleware('auth')->name('admins.removeadmin');
 
-Route::delete('/admins/destroy/{user}','AdminController@destroy')->middleware('admin')->name('admins.destroy');
-Route::delete('/admins/delete/{event}','AdminController@deleteEvent')->middleware('admin')->name('admins.delete');
+Route::delete('/admins/destroy/{user}','AdminController@destroy')->middleware('admin')->middleware('auth')->name('admins.destroy');
+Route::delete('/admins/delete/{event}','AdminController@deleteEvent')->middleware('admin')->middleware('auth')->name('admins.delete');
